refactor: card.jsx에 있던 컴포넌트들을 별도 파일로 분리
